+++ src/crab/type_domain.hpp
@@ -19,24 +19,6 @@ enum class region {
 	T_SHARED
 };
 
-inline std::string get_reg_ptr(const region& r) {
-    switch (r) {
-        case region::T_CTX:
-            return "ctx_p";
-        case region::T_STACK:
-            return "stack_p";
-        case region::T_PACKET:
-            return "packet_p";
-        default:
-            return "shared_p";
-    }
-}
-
-inline std::ostream& operator<<(std::ostream& o, const region& t) {
-    o << static_cast<std::underlying_type<region>::type>(t);
-    return o;
-}
-
 struct ptr_no_off_t {
     region r;
 
@@ -45,21 +27,11 @@ struct ptr_no_off_t {
     ptr_no_off_t(ptr_no_off_t &&) = default;
     ptr_no_off_t &operator=(const ptr_no_off_t &) = default;
     ptr_no_off_t &operator=(ptr_no_off_t &&) = default;
-
     ptr_no_off_t(region _r) : r(_r) {}
 
-    friend std::ostream& operator<<(std::ostream& o, const ptr_no_off_t& p) {
-        return o << get_reg_ptr(p.r);
-    }
-  
-    // temporarily make operators friend functions in order to avoid duplicate symbol errors
-    friend bool operator==(const ptr_no_off_t& p1, const ptr_no_off_t& p2) {
-        return (p1.r == p2.r);
-    }
-
-    friend bool operator!=(const ptr_no_off_t& p1, const ptr_no_off_t& p2) {
-        return !(p1 == p2);
-    }
+    friend std::ostream& operator<<(std::ostream& o, const ptr_no_off_t& p);
+    friend bool operator==(const ptr_no_off_t& p1, const ptr_no_off_t& p2);
+    friend bool operator!=(const ptr_no_off_t& p1, const ptr_no_off_t& p2);
 };
 
 struct ptr_with_off_t {
@@ -71,22 +43,11 @@ struct ptr_with_off_t {
     ptr_with_off_t(ptr_with_off_t &&) = default;
     ptr_with_off_t &operator=(const ptr_with_off_t &) = default;
     ptr_with_off_t &operator=(ptr_with_off_t &&) = default;
-
     ptr_with_off_t(region _r, int _off) : r(_r), offset(_off) {}
 
-    friend std::ostream& operator<<(std::ostream& o, const ptr_with_off_t& p) {
-        o << get_reg_ptr(p.r) << "<" << p.offset << ">";
-        return o;
-    }
-
-    // temporarily make operators friend functions in order to avoid duplicate symbol errors
-    friend bool operator==(const ptr_with_off_t& p1, const ptr_with_off_t& p2) {
-        return (p1.r == p2.r && p1.offset == p2.offset);
-    }
-
-    friend bool operator!=(const ptr_with_off_t& p1, const ptr_with_off_t& p2) {
-        return !(p1 == p2);
-    }
+    friend std::ostream& operator<<(std::ostream& o, const ptr_with_off_t& p);
+    friend bool operator==(const ptr_with_off_t& p1, const ptr_with_off_t& p2);
+    friend bool operator!=(const ptr_with_off_t& p1, const ptr_with_off_t& p2);
 };
 
 using ptr_t = std::variant<ptr_no_off_t, ptr_with_off_t>;
@@ -97,57 +58,10 @@ struct reg_with_loc_t {
     std::pair<label_t, uint32_t> loc;
 
     reg_with_loc_t(register_t _r, const label_t& l, uint32_t loc_instr) : r(_r), loc(std::make_pair(l, loc_instr)) {}
-
-    bool operator==(const reg_with_loc_t& other) const {
-        return (r == other.r && loc == other.loc);
-    }
-
-    std::size_t hash() const {
-        // Similar to boost::hash_combine
-        using std::hash;
-
-        std::size_t seed = hash<register_t>()(r);
-        seed ^= hash<int>()(loc.first.from) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
-        seed ^= hash<int>()(loc.second) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
-
-        return seed;
-    }
-
-    friend std::ostream& operator<<(std::ostream& o, const reg_with_loc_t& reg) {
-        o << "r" << static_cast<unsigned int>(reg.r) << "@" << reg.loc.second << " in " << reg.loc.first;
-        return o;
-    }
+    bool operator==(const reg_with_loc_t& other) const;
+    std::size_t hash() const;
+    friend std::ostream& operator<<(std::ostream& o, const reg_with_loc_t& reg);
 };
-}
-
-
-namespace std {
-    template <>
-    struct std::hash<crab::reg_with_loc_t> {
-        std::size_t operator()(const crab::reg_with_loc_t& reg) const { return reg.hash(); }
-    };
-
-    // does not seem to work for me
-    template <>
-    struct std::equal_to<crab::ptr_t> {
-        constexpr bool operator()(const crab::ptr_t& p1, const crab::ptr_t& p2) const {
-            if (p1.index() != p2.index()) return false;
-            if (std::holds_alternative<crab::ptr_no_off_t>(p1)) {
-                auto ptr_no_off1 = std::get<crab::ptr_no_off_t>(p1);
-                auto ptr_no_off2 = std::get<crab::ptr_no_off_t>(p2);
-                return (ptr_no_off1.r == ptr_no_off2.r);
-            }
-            else {
-                auto ptr_with_off1 = std::get<crab::ptr_with_off_t>(p1);
-                auto ptr_with_off2 = std::get<crab::ptr_with_off_t>(p2);
-                return (ptr_with_off1.r == ptr_with_off2.r && ptr_with_off1.offset == ptr_with_off2.offset);
-            }
-        }
-    };
-}
-
-
-namespace crab {
 
 class ctx_t {
     using offset_to_ptr_no_off_t = std::unordered_map<int, ptr_no_off_t>;
@@ -155,28 +69,9 @@ class ctx_t {
     offset_to_ptr_no_off_t m_packet_ptrs;
 
   public:
-    ctx_t(const ebpf_context_descriptor_t* desc)
-    {
-        if (desc->data != -1)
-            m_packet_ptrs[desc->data] = crab::ptr_no_off_t(crab::region::T_PACKET);
-        if (desc->end != -1)
-            m_packet_ptrs[desc->end] = crab::ptr_no_off_t(crab::region::T_PACKET);
-    }
-
-    std::optional<ptr_no_off_t> find(int key) const {
-        auto it = m_packet_ptrs.find(key);
-        if (it == m_packet_ptrs.end()) return {};
-        return it->second;
-    }
-
-    friend std::ostream& operator<<(std::ostream& o, const ctx_t& _ctx) {
-
-        o << "type of context: " << (_ctx.m_packet_ptrs.empty() ? "_|_" : "") << "\n";
-        for (const auto& it : _ctx.m_packet_ptrs) {
-            o << "  stores at " << it.first << ": " << it.second << "\n";
-        }
-        return o;
-    }
+    ctx_t(const ebpf_context_descriptor_t* desc);
+    std::optional<ptr_no_off_t> find(int key) const;
+    friend std::ostream& operator<<(std::ostream& o, const ctx_t& _ctx);
 };
 
 class stack_t {
@@ -190,53 +85,16 @@ class stack_t {
     stack_t(offset_to_ptr_t && ptrs, bool is_bottom)
     : m_ptrs(std::move(ptrs)) , m_is_bottom(is_bottom) {}
     
-    stack_t operator|(const stack_t& other) const {
-        if (is_bottom() || other.is_top()) {
-            return other;
-        } else if (other.is_bottom() || is_top()) {
-            return *this;
-        }
-        offset_to_ptr_t out_ptrs;
-        for (auto const&kv: m_ptrs) {
-            auto it = other.find(kv.first);
-            if (it && kv.second == it.value())
-                out_ptrs.insert(kv);
-        }
-        return stack_t(std::move(out_ptrs), false);
-    }
-
-    void set_to_bottom() {
-        m_ptrs.clear();
-        m_is_bottom = true;
-    }
-
-    void set_to_top() {
-        m_ptrs.clear();
-        m_is_bottom = false;
-    }
-
-    static stack_t bottom() { return stack_t(true); }
-
-    static stack_t top() { return stack_t(false); }
-
-    bool is_bottom() const { return m_is_bottom; }
-
-    bool is_top() const {
-        if (m_is_bottom)
-            return false;
-        return m_ptrs.empty();
-    }
-
+    stack_t operator|(const stack_t& other) const;
+    void set_to_bottom();
+    void set_to_top();
+    static stack_t bottom();
+    static stack_t top();
+    bool is_bottom() const;
+    bool is_top() const;
     const offset_to_ptr_t &get_ptrs() { return m_ptrs; }
-
-    void insert(int key, ptr_t value) { m_ptrs.insert(std::make_pair(key, value)); }
-
-    std::optional<ptr_t> find(int key) const {
-        auto it = m_ptrs.find(key);
-        if (it == m_ptrs.end()) return {};
-        return it->second;
-    }
-
+    void insert(int key, ptr_t value);
+    std::optional<ptr_t> find(int key) const;
     friend std::ostream& operator<<(std::ostream& o, const stack_t& st);
 };
 
@@ -253,74 +111,17 @@ class register_types_t {
     explicit register_types_t(live_registers_t&& vars, std::shared_ptr<global_type_env_t> all_types, bool is_bottom = false)
         : m_vars(std::move(vars)), m_all_types(all_types), m_is_bottom(is_bottom) {}
 
-    register_types_t operator|(const register_types_t& other) const {
-        if (is_bottom() || other.is_top()) {
-            return other;
-        } else if (other.is_bottom() || is_top()) {
-            return *this;
-        }
-        live_registers_t out_vars;
-        for (size_t i = 0; i < m_vars.size(); i++) {
-            if (m_vars[i] == nullptr) continue;
-            auto it1 = find(*(m_vars[i]));
-            auto it2 = other.find(*(other.m_vars[i]));
-            if (it1 && it2 && it1.value() == it2.value()) {
-                out_vars[i] = m_vars[i];
-            }
-        }
-
-        return register_types_t(std::move(out_vars), m_all_types, false);
-    }
-
-    void operator-=(register_t var) {
-        if (is_bottom()) {
-            return;
-        }
-        m_vars[var] = nullptr;
-    }
-
-    void set_to_bottom() {
-        m_vars = live_registers_t{nullptr};
-        m_is_bottom = true;
-    }
-
-    void set_to_top() {
-        m_vars = live_registers_t{nullptr};
-        m_is_bottom = false;
-    }
-
-    bool is_bottom() const { return m_is_bottom; }
-
-    bool is_top() const {
-        if (m_is_bottom) { return false; }
-        if (m_all_types == nullptr) return true;
-        for (auto it : m_vars) {
-            if (it != nullptr) return false;
-        }
-        return true;
-    }
-
-    void insert(register_t reg, const reg_with_loc_t& reg_with_loc, const ptr_t& type) {
-        (*m_all_types)[reg_with_loc] = type;
-        m_vars[reg] = std::make_shared<reg_with_loc_t>(reg_with_loc);
-    }
-
-    std::optional<ptr_t> find(reg_with_loc_t reg) const {
-        auto it = m_all_types->find(reg);
-        if (it == m_all_types->end()) return {};
-        return it->second;
-    }
-
-    std::optional<ptr_t> find(register_t key) const {
-        if (m_vars[key] == nullptr) return {};
-        const reg_with_loc_t& reg = *(m_vars[key]);
-        return find(reg);
-    }
-
+    register_types_t operator|(const register_types_t& other) const;
+    void operator-=(register_t var);
+    void set_to_bottom();
+    void set_to_top();
+    bool is_bottom() const;
+    bool is_top() const;
+    void insert(register_t reg, const reg_with_loc_t& reg_with_loc, const ptr_t& type);
+    std::optional<ptr_t> find(reg_with_loc_t reg) const;
+    std::optional<ptr_t> find(register_t key) const;
     const live_registers_t &get_vars() { return m_vars; }
-
     friend std::ostream& operator<<(std::ostream& o, const register_types_t& p);
-
 };
 
 }
@@ -335,83 +136,60 @@ class type_domain_t final {
 
   public:
 
-  type_domain_t() : m_label(label_t::entry) {}
-  type_domain_t(type_domain_t&& o) = default;
-  type_domain_t(const type_domain_t& o) = default;
-  type_domain_t& operator=(type_domain_t&& o) = default;
-  type_domain_t& operator=(const type_domain_t& o) = default;
-  type_domain_t(crab::register_types_t&& _types, crab::stack_t&& _st, const label_t& _l, std::shared_ptr<crab::ctx_t> _ctx)
+    type_domain_t() : m_label(label_t::entry) {}
+    type_domain_t(type_domain_t&& o) = default;
+    type_domain_t(const type_domain_t& o) = default;
+    type_domain_t& operator=(type_domain_t&& o) = default;
+    type_domain_t& operator=(const type_domain_t& o) = default;
+    type_domain_t(crab::register_types_t&& _types, crab::stack_t&& _st, const label_t& _l, std::shared_ptr<crab::ctx_t> _ctx)
             : m_stack(std::move(_st)), m_types(std::move(_types)), m_ctx(_ctx), m_label(_l) {}
-  // eBPF initialization: R1 points to ctx, R10 to stack, etc.
-  static type_domain_t setup_entry();
-  // bottom/top
-  static type_domain_t bottom();
-  void set_to_top();
-  void set_to_bottom();
-  bool is_bottom() const;
-  bool is_top() const;
-  // inclusion
-  bool operator<=(const type_domain_t& other) const;
-  // join
-  void operator|=(const type_domain_t& abs);
-  void operator|=(type_domain_t&& abs);
-  type_domain_t operator|(const type_domain_t& other) const;
-  type_domain_t operator|(type_domain_t&& abs) const;
-  // meet
-  type_domain_t operator&(const type_domain_t& other) const;
-  // widening
-  type_domain_t widen(const type_domain_t& other) const;
-  // narrowing
-  type_domain_t narrow(const type_domain_t& other) const;
-  //forget
-  void operator-=(variable_t var);
+    // eBPF initialization: R1 points to ctx, R10 to stack, etc.
+    static type_domain_t setup_entry();
+    // bottom/top
+    static type_domain_t bottom();
+    void set_to_top();
+    void set_to_bottom();
+    bool is_bottom() const;
+    bool is_top() const;
+    // inclusion
+    bool operator<=(const type_domain_t& other) const;
+    // join
+    void operator|=(const type_domain_t& abs);
+    void operator|=(type_domain_t&& abs);
+    type_domain_t operator|(const type_domain_t& other) const;
+    type_domain_t operator|(type_domain_t&& abs) const;
+    // meet
+    type_domain_t operator&(const type_domain_t& other) const;
+    // widening
+    type_domain_t widen(const type_domain_t& other) const;
+    // narrowing
+    type_domain_t narrow(const type_domain_t& other) const;
+    //forget
+    void operator-=(variable_t var);
 
-  //// abstract transformers
-  void operator()(const Undefined &);
-  void operator()(const Bin &);
-  void operator()(const Un &) ;
-  void operator()(const LoadMapFd &);
-  void operator()(const Call &);
-  void operator()(const Exit &);
-  void operator()(const Jmp &);
-  void operator()(const Mem &);
-  void operator()(const Packet &);
-  void operator()(const LockAdd &);
-  void operator()(const Assume &);
-  void operator()(const Assert &);
-  void operator()(const basic_block_t& bb, bool check_termination) {
-      m_curr_pos = 0;
-      m_label = bb.label();
-      std::cout << m_label << ":\n";
-      for (const Instruction& statement : bb) {
-        m_curr_pos++;
-        std::visit(*this, statement);
-    }
-    auto [it, et] = bb.next_blocks();
-    if (it != et) {
-        std::cout << "  "
-          << "goto ";
-        for (; it != et;) {
-            std::cout << *it;
-            ++it;
-            if (it == et) {
-                std::cout << ";";
-            } else {
-                std::cout << ",";
-            }
-        }
-    }
-    std::cout << "\n\n";
-  }
-  void write(std::ostream& os) const;
-  std::string domain_name() const;
-  int get_instruction_count_upper_bound();
-  string_invariant to_set();
-  void set_require_check(check_require_func_t f);
+    //// abstract transformers
+    void operator()(const Undefined &);
+    void operator()(const Bin &);
+    void operator()(const Un &) ;
+    void operator()(const LoadMapFd &);
+    void operator()(const Call &);
+    void operator()(const Exit &);
+    void operator()(const Jmp &);
+    void operator()(const Mem &);
+    void operator()(const Packet &);
+    void operator()(const LockAdd &);
+    void operator()(const Assume &);
+    void operator()(const Assert &);
+    void operator()(const basic_block_t& bb, bool check_termination);
+    void write(std::ostream& os) const;
+    std::string domain_name() const;
+    int get_instruction_count_upper_bound();
+    string_invariant to_set();
+    void set_require_check(check_require_func_t f);
 
   private:
 
-  void do_load(const Mem&, const Reg&);
-  void do_mem_store(const Mem&, const Reg&);
+    void do_load(const Mem&, const Reg&);
+    void do_mem_store(const Mem&, const Reg&);
 
 }; // end type_domain_t
